add class InvalidDataException

+++ src/main/java/org/HMB/ClientService.java
@@ -1,18 +1,18 @@
 package org.HMB;
 
 import org.HMB.entities.Client;
+import org.HMB.exception.InvalidDataException;
 
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
 public class ClientService {
-    private PreparedStatement createSt;
-    private PreparedStatement getByNameSt;
-    private PreparedStatement getByIdSt;
-    private PreparedStatement setNameSt;
-    private PreparedStatement deleteByIdSt;
-    private PreparedStatement listAllSt;
+    private final PreparedStatement createSt;
+    private final PreparedStatement getByIdSt;
+    private final PreparedStatement setNameSt;
+    private final PreparedStatement deleteByIdSt;
+    private final PreparedStatement listAllSt;
 
     public ClientService(Connection connection){
         try {
@@ -26,23 +26,29 @@ public class ClientService {
         }
     }
 
-    public long create(String name) {
+    public long create(String name) throws InvalidDataException {
         Long id = null;
+        if (name.length() < 2 || name.length() > 100) {
+            throw new InvalidDataException("is not validate!");
+        }
         try {
             createSt.setString(1, name);
             createSt.executeUpdate();
             ResultSet resultSet = createSt.getGeneratedKeys();
-            if(resultSet.next()) {
-                id = resultSet.getLong(1);
-            }
-
+                if (resultSet.next()) {
+                    id = resultSet.getLong(1);
+                }
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
         return id;
     }
 
-    public String getById(long id){
+    public String getById(long id) throws InvalidDataException {
+
+        if (id <= 0) {
+            throw new InvalidDataException("ID not validate!");
+        }
         try {
             getByIdSt.setLong(1, id);
             ResultSet resultSet = getByIdSt.executeQuery();
@@ -55,7 +61,11 @@ public class ClientService {
             throw new RuntimeException(e);
         }
     }
-    public void setName(long id, String name){
+    public void setName(long id, String name) throws InvalidDataException {
+
+        if (id <= 0 || name.length() < 2 || name.length() > 100) {
+            throw new InvalidDataException("data not validate!");
+        }
         try {
             setNameSt.setString(1, name);
             setNameSt.setLong(2, id);
@@ -63,9 +73,13 @@ public class ClientService {
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-
     }
-    public void deleteById(long id){
+    public void deleteById(long id) throws InvalidDataException {
+
+        if (id <= 0) {
+            throw new InvalidDataException("ID not validate!");
+        }
+
         try {
             deleteByIdSt.setLong(1, id);
             deleteByIdSt.executeUpdate();
